pyproject.toml
[project]
name = "pilotx-agent"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sangram Reddy", email = "reddy.horcrux@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "firebase-admin>=7.1.0",
    "a2a-sdk[grpc,sql]",
    "google-adk>=1.18.0",
    "litellm>=1.79.3",
    "tavily-python>=0.7.12",
    "tenacity>=9.1.2",
    "toolbox-core>=0.5.2",
    "yamale>=6.0.0",
    "rich>=14.2.0",
]

[project.scripts]
pilotx-agent = "pilotx_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.10.0",
    "build>=1.2.2.post1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.4",
    "setuptools>=80.9.0",
    "jupyter>=1.1.1",
    "streamlit>=1.45.1",
    "testcontainers>=3.7.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.uv]
environments = [
    "sys_platform == 'darwin'",
    "sys_platform == 'linux'",
]

[tool.hatch.version]
path = "src/pilotx_agent/__init__.py"
